requirements.txt
streamlit~=1.27.2
requests~=2.31.0
pandas~=2.0.3
feedparser~=6.0.10
bs4~=4.12.2
beautifulsoup4~=4.12.2
python-dateutil~=2.8.2
numpy~=1.26.0
nltk~=3.8.1
scikit-learn~=1.3.0
arxiv~=1.4.8
seaborn~=0.12.2
plotly~=5.17.0